pyproject.toml
[tool.poetry]
name = "navi-api"
version = "0.1.0"
description = ""
authors = ["suk-6 <me@suk.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
black = "^24.10.0"
pydantic-settings = "^2.6.1"
requests = "^2.32.3"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
